requirements.txt
# Framework
django==2.2.9
django-environ==0.4.5
django-redis==4.11.0
django-redis-sessions==0.6.2
djangorestframework==3.9.4
rest-condition==1.0.3
argon2-cffi==19.2.0  # password hasher

# PostgreSQL
psycopg2==2.8.6
psycopg2-binary==2.8.6

# Gunicorn
gunicorn==20.0.4

# Encryption
pycryptodome==3.9.7
numpy==1.16.3

# QRCode
qrcode==6.1
Pillow==8.2.0

# Celery
celery==4.4.0
flower==0.9.3

# OTP
pyotp==2.2.7

# Telegram
python-telegram-bot==12.7

# Http request
requests==2.23.0

# Testing
coverage==5.1
freezegun==0.3.15

# AWS
boto3==1.13.1

# Excel
XlsxWriter==1.3.7
pip~=21.1.1
locust~=1.5.3